+++ src/physical-layer/channel-receive/channel-receive.factory.js
@@ -7,7 +7,7 @@
 
     _ChannelReceive.$inject = [
         'PhysicalLayer.Audio',
-        'PhysicalLayer.CarrierRecoveryBuilder',
+        'Common.CarrierRecoveryBuilder',
         'Common.MathUtil'
     ];